requirements.txt
mysql-connector-python==8.3.0
langchain
langchain-community
google-generativeai
streamlit
langchain-google-genai
python-dotenv
chromadb==0.3.29
pypdf
PyPDF2
pandas
pysqlite3-binary
SpeechRecognition
pyttsx3
streamlit-webrtc